+++ src/Rdd.Domain/Models/Querying/IStringConverter.cs
@@ -1,9 +1,6 @@
 ﻿using System;
-using System.Linq;
 using System.Collections;
 using System.Collections.Generic;
-using System.Linq.Expressions;
-using System.Reflection;
 
 namespace Rdd.Domain.Models.Querying
 {
@@ -14,29 +11,5 @@ namespace Rdd.Domain.Models.Querying
 
         IList ConvertValues(Type type, IEnumerable<string> values);
         object ConvertTo(Type type, string input);
-    }
-
-    public static class ExpressionExtension
-    {
-        public static Expression ExtractExpression<T>(this T value)
-        {
-            if (typeof(T) == typeof(object))
-            {
-                throw new NotSupportedException("Only use typed values");
-            }
-            return ((Expression<Func<T>>)(() => value)).Body;
-        }
-
-        public static Expression ExtractTypedExpression(this object value, Type type)
-        {
-            var method = typeof(ExpressionExtension)
-                .GetMethod(nameof(ExtractExpression),BindingFlags.Public | BindingFlags.Static);
-            if (method == null)
-            {
-                throw new NotSupportedException();
-            }
-            var generic = method.MakeGenericMethod(type);
-            return (Expression)generic.Invoke(null, new object[] { value });
-        }
     }
 }
