+++ src/Rdd.Infra/Helpers/WebFilterConverter.cs
@@ -8,7 +8,6 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
-using Rdd.Domain.Models.Querying;
 
 namespace Rdd.Infra.Helpers
 {
